+++ packages/react-tenant-selector/src/hooks/useClusterSelector.ts
@@ -60,11 +60,9 @@ const useClusterSelector = (appName: string) => {
 
   const isStaging = hostname.includes('.staging.');
 
-  const initialCluster = localStorage.getItem('initialCluster');
-
   const onClusterSelected = (newTenant: string, cluster: string) => {
     localStorage.setItem('initialTenant', newTenant);
-    localStorage.setItem('initialCluster', cluster);
+
     setRedirectingToCluster(true);
     const { hash, search } = window.location;
     const url = [
@@ -107,7 +105,6 @@ const useClusterSelector = (appName: string) => {
     onClusterSelected,
     redirectingToCluster,
     validatingCluster,
-    initialCluster,
   };
 };