+++ packages/react-tenant-selector/src/hooks/useTenantSelector.ts
@@ -9,7 +9,6 @@ const useTenantSelector = (appName: string, appsApiBaseUrl: string) => {
 
   const onTenantSelected = (newTenant: string) => {
     localStorage.setItem('initialTenant', newTenant);
-    localStorage.removeItem('initialCluster');
     setRedirecting(true);
     const { hash, search } = window.location;
     const url = [`/${newTenant}`, search, hash].filter(Boolean).join('');